[#932] Moves form componentes to separate directories
with @deniscostadsc
