+++ web-ui/src/account_recovery/admin_recovery_code_form/admin_recovery_code_form.js
@@ -20,36 +20,35 @@ import { translate } from 'react-i18next';
 
 import InputField from 'src/common/input_field/input_field';
 import SubmitButton from 'src/common/submit_button/submit_button';
-import BackLink from 'src/common/back_link/back_link';
 
-export const NewPasswordForm = ({ t, previous }) => (
-  <form className='account-recovery-form new-password'>
+import './admin_recovery_code_form.scss';
+
+export const AdminRecoveryCodeForm = ({ t, next }) => (
+  <form className='account-recovery-form admin-code' onSubmit={next}>
     <img
       className='account-recovery-progress'
-      src='/public/images/account-recovery/step_3.svg'
-      alt={t('account-recovery.new-password.image-description')}
+      src='/public/images/account-recovery/step_1.svg'
+      alt={t('account-recovery.admin-form.image-description')}
     />
-    <h1>{t('account-recovery.new-password-form.title')}</h1>
-    <InputField
-      type='password' name='new-password'
-      label={t('account-recovery.new-password-form.input-label1')}
+    <h1>{t('account-recovery.admin-form.title')}</h1>
+    <img
+      className='admin-code-image'
+      src='/public/images/account-recovery/admins_contact.svg'
+      alt=''
     />
-    <InputField
-      type='password' name='confirm-password'
-      label={t('account-recovery.new-password-form.input-label2')}
-    />
-    <SubmitButton buttonText={t('account-recovery.new-password-form.button')} />
-    <BackLink
-      text={t('account-recovery.back')}
-      onClick={previous} onKeyDown={previous}
-      role='button'
-    />
+    <ul>
+      <li>{t('account-recovery.admin-form.tip1')}</li>
+      <li>{t('account-recovery.admin-form.tip2')}</li>
+      <li>{t('account-recovery.admin-form.tip3')}</li>
+    </ul>
+    <InputField name='admin-code' label={t('account-recovery.admin-form.input-label')} />
+    <SubmitButton buttonText={t('account-recovery.admin-form.button')} />
   </form>
 );
 
-NewPasswordForm.propTypes = {
+AdminRecoveryCodeForm.propTypes = {
   t: React.PropTypes.func.isRequired,
-  previous: React.PropTypes.func.isRequired
+  next: React.PropTypes.func.isRequired
 };
 
-export default translate('', { wait: true })(NewPasswordForm);
+export default translate('', { wait: true })(AdminRecoveryCodeForm);
