--- web-ui/src/account_recovery/forms/forms.scss
@@ -1,92 +1,0 @@
-/*
- * Copyright (c) 2017 ThoughtWorks, Inc.
- *
- * Pixelated is free software: you can redistribute it and/or modify
- * it under the terms of the GNU Affero General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * Pixelated is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU Affero General Public License for more details.
- *
- * You should have received a copy of the GNU Affero General Public License
- * along with Pixelated. If not, see <http://www.gnu.org/licenses/>.
- */
-
-.account-recovery-form {
-  display: flex;
-  flex-direction: column;
-
-  img {
-    margin: 1em 0;
-    align-self: center;
-  }
-
-  .user-code-form-content {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-  }
-
-  .account-recovery-progress {
-    width: 100%;
-  }
-
-  .admin-code-image {
-    height: 2.7em;
-  }
-
-  .user-code-image {
-    height: 4em;
-  }
-
-  .input-field-group {
-    margin-top: 0;
-  }
-}
-
-.new-password {
-  .input-field-group:first-of-type {
-    margin-bottom: 0;
-  }
-}
-
-@media only screen and (min-width : 500px) {
-  .account-recovery-form {
-    align-items: center;
-
-    .account-recovery-progress, h1 {
-      width: 80%;
-    }
-
-    .user-code-form-content {
-      flex-direction: row;
-      width: 80%;
-
-      img {
-        margin: 1.6em;
-      }
-    }
-  }
-}
-
-@media only screen and (min-width : 960px) {
-  .account-recovery-form {
-    .account-recovery-progress {
-      width: 80%;
-      margin-top: 0;
-    }
-
-    h1 {
-      max-width: 80%;
-      width: auto;
-    }
-
-    .input-field-group, .submit-button {
-      width: 60%;
-      align-self: center;
-    }
-  }
-}
