+++ web-ui/src/account_recovery/new_password_form/new_password_form.js
@@ -22,26 +22,25 @@ import InputField from 'src/common/input_field/input_field';
 import SubmitButton from 'src/common/submit_button/submit_button';
 import BackLink from 'src/common/back_link/back_link';
 
-import './forms.scss';
+import './new_password_form.scss';
 
-export const UserRecoveryCodeForm = ({ t, previous, next }) => (
-  <form className='account-recovery-form user-code' onSubmit={next}>
+export const NewPasswordForm = ({ t, previous }) => (
+  <form className='account-recovery-form new-password'>
     <img
       className='account-recovery-progress'
-      src='/public/images/account-recovery/step_2.svg'
-      alt={t('account-recovery.user-form.image-description')}
+      src='/public/images/account-recovery/step_3.svg'
+      alt={t('account-recovery.new-password.image-description')}
     />
-    <h1>{t('account-recovery.user-form.title')}</h1>
-    <div className='user-code-form-content'>
-      <img
-        className='user-codes-image'
-        src='/public/images/account-recovery/codes.svg'
-        alt=''
-      />
-      <p>{t('account-recovery.user-form.description')}</p>
-    </div>
-    <InputField name='admin-code' label={t('account-recovery.user-form.input-label')} />
-    <SubmitButton buttonText={t('account-recovery.user-form.button')} />
+    <h1>{t('account-recovery.new-password-form.title')}</h1>
+    <InputField
+      type='password' name='new-password'
+      label={t('account-recovery.new-password-form.input-label1')}
+    />
+    <InputField
+      type='password' name='confirm-password'
+      label={t('account-recovery.new-password-form.input-label2')}
+    />
+    <SubmitButton buttonText={t('account-recovery.new-password-form.button')} />
     <BackLink
       text={t('account-recovery.back')}
       onClick={previous} onKeyDown={previous}
@@ -50,10 +49,9 @@ export const UserRecoveryCodeForm = ({ t, previous, next }) => (
   </form>
 );
 
-UserRecoveryCodeForm.propTypes = {
+NewPasswordForm.propTypes = {
   t: React.PropTypes.func.isRequired,
-  previous: React.PropTypes.func.isRequired,
-  next: React.PropTypes.func.isRequired
+  previous: React.PropTypes.func.isRequired
 };
 
-export default translate('', { wait: true })(UserRecoveryCodeForm);
+export default translate('', { wait: true })(NewPasswordForm);
