+++ web-ui/src/account_recovery/user_recovery_code_form/user_recovery_code_form.scss
@@ -1,0 +1,39 @@
+/*
+ * Copyright (c) 2017 ThoughtWorks, Inc.
+ *
+ * Pixelated is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Affero General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * Pixelated is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU Affero General Public License for more details.
+ *
+ * You should have received a copy of the GNU Affero General Public License
+ * along with Pixelated. If not, see <http://www.gnu.org/licenses/>.
+ */
+
+.user-code-form-content {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+}
+
+.user-code-image {
+  margin: 1em 0;
+  align-self: center;
+  height: 4em;
+}
+
+@media only screen and (min-width : 500px) {
+  .user-code-form-content {
+    flex-direction: row;
+    width: 80%;
+  }
+
+  .user-code-image {
+    margin-right: 1.6em;
+  }
+}
